Make upper and lower limits of fan speed configurable
Although any value between 0% and 100% can be selected as fan speed
very low or very high values were clipped to 20% and 80% respectively.
This happens because the GPU vendors limit the lowest and highest
fan speeds in the VGA BIOS. In order to somehow mimic this behaviour
the above clipping was used. However the fan speed limits are
totally arbitrary and depend on the choices of the vendor. Therefore
there is no point in hard-coding the limits to 20% and 80%. Although
this conservative approach still remains as default the limits are
now configurable through a new command line switch, -l or --limits.