Proper NvidiaControl error propagation and handling
In order to avoid crashing during initialization and deinitialization
proper error propagation must be made when instantiating an NvidiaControl
struct. Since NvidiaControl::init() errors were not handled they would
cause a panic in the main thread. As far as NvidiaControl::deinit() is
concerned, as it is called during drop() it is not possible to handle
the error properly, so it is dropped. Hopefully we will never reach
drop() if the NvidiaControl::init() call is not successful as the
NvidiaControl struct is only returned upon successful library load.