More robust nvapi DLL loading on windows
Instead of using a hardcoded &str for the nvapi DLL on windows we derive
the correct path from the %SystemRoot% environment variable. This way
some redundant code is eliminated and in the off chance that Windows is
not installed in C:\ the DLL is properly loaded. In case %SystemRoot% is
not defined the default "C:\\Windows" is used.