Use FFI instead of C code to bind into XNVCtrl
Up to now nvfancontrol in Linux used a small C glue library to interface
with XNVCtrl. This commit removes all C code and uses the Rust FFI
facility to directly bind with Xlib and XNVCtrl as we do with NVAPI in
Windows. Therefore the X11 and NVCtrl development headers are no longer
required when building the project.