[Windows] Implement partial support for cooler policies
In contrast to the XNVCtrl API where the cooler fan can be either
automatically or manually managed, in Windows the cooler fan is
subject to cooler policies. Currently there are 6 different
policies, Manual (2^0), Performance (2^1), Discrete Temperature
Steps (2^2), Continuous Hardware Curve (2^3), Continuous Software
Curve (2^4) and the meta-policy Default (2^5). For the purpose
of nvfancontrol only Manual and Default is used for
NVCtrlFanControlState::Manual and ::Auto respectively.