[Windows] Allocate a single GPU handle for NvidiaControl
Instead of instantiating a GPU handle every time a NvAPI related
function is called create only a single handle as part of the
NvidiaControl struct. This way a lot of the boilerplate code for
Windows is eliminated. This _could_ create an issue when multiple
threads are trying to access the handle however this also holds
true for the NvidiaControl struct itself anyway. If a need for
multithreaded access arises some synchronisation primitives can be
employed.