Add a minimal TCP server to facilitate third party queries
An optional barebones TCP server can be started with the -t option. By
passing an additional optional parameter the port can be customised from
the default 12125. This can be useful for programs that want to know the
temperature and fan status without parsing stdout. Upon connection the
server immediately sends the current data and closes the connection. The
server is implemented as a single thread so it probably can't handle many
connections as it is blocking and all the requests are handled within the
thread. The thread is terminated when the AtomicBool that handles the main
loop is set to false after it exits.  Before handling any request the
thread checks for the value of this AtomicBool and if it's false the thread
terminates as well. The downside of this approach is that there is no way
to notify the thread that the AtomicBool is set other than "forcing" it to
handle a request. The ideal way would probably be to use some timeouts in
the TCP handling thread however this would require the use of an async
framework which will complicate the program significantly for negligible
benefits.