Add an additional nvctrl function to return the number of available GPUs
Function `gpu_count` has been added to the NvFanController trait to
return the number of available GPUs. Since this information is permanent
during the lifetime of the program it is stored in the `_gpu_count` field
of the `NvidiaControl` struct. The number of GPUs is determined during
the initialisation of `NvidiaControl`. An additional function named
`check_gpu_id` has also been added for both Unix and Windows which
checks if the specified GPU id is valid, ie. if the GPU id is not greater
than the number of available GPUs in the system. `check_gpu_id` can be
used in conjuction the try!() macro to early exit a function that requires
access to a specified GPU if the supplied id is invalid.