Apply fan curve to all coolers of the GPU; currently only for Unix.
NvidiaControl now holds a list of all available GPUs with their associated
coolers. For XNVCtrl coolers are numbered globally rather than per GPU so
cooler ids must be tracked in order to be used later.

Main program now applies the curve to all coolers of the GPU and the GPU
is selectable with the `-g` switch. This will probably evolve into a list
so that different curves can be applied per GPU rather than be limited to
a single adpater.

An additional switch `-p` has been added to list all adapters with their
associated coolers.