Use Cow<Vec> instead of Vec for returning the GPU coolers.
This is a necessary change because the Vec with the cooler ids can either
be preallocated on NvFanController instantiation (on Unix) or can be
generated during function call (on Windows; probably). When it's already
allocated the function will only return a reference since that's all
is needed and the actual structure does not need to change after program
startup.