Initial code for multiple cooler support on Windows.
Although this commit builds cleanly it has not been tested thoroughly.
There will DEFINITELY be bugs! There are two issues that are relatively.
One is that RPM readings are only provided by `NvAPI_GPU_GetTachReading`
which does not allow indexing on the coolers. So the RPM will always be
the same for all coolers or even totally wrong. This is not something that
can be solved without alterations to NvAPI, although it is quite probable
that it has already been resolved in the NDA version of the API. But we
can't have that so there isn't much we can do. The second is wrt the
cooler policies. In contrast to linux where the GPU fan policy is global
per GPU, NVAPI allows having a per cooler policy for the fans. As
`nvfancontrol` deals with fan control rather than providing an
XNVCtrl/NVAPI abstraction it will assume that fan policy is per GPU on
Windows as well. This is not ideal but I believe it makes sense that if
one needs manual control would expect it to apply on all fans.