Add support for TOML configuration files
Further to the efforts to add multi-GPU support and in order to
support individual curves per GPU alterations to the configuration
files have been made. As a flat 2-col configuration file can not
support this kind of setup we now support TOML configuration files.
A list of `[[gpus]]` block can now be provided in the TOML file.
Structure of said file is as follows

```
[[gpus]]
id = 0
enabled = true
points = [[temp, speed], [temp, speed], [...]]
```

Each block corresponds to a different GPU so the id of each MUST be
different. For compatibility reasons existing 2-col "legacy"
configuration files are still parseable and the GPU upon which they
operate is either 0 or the one specified in the command line.

Due to the dependence of `toml` crate on `serde` we now fix the versions
of both to 0.5 and 1.0 respectively.