Properly load global configuration file from /etc/xdg on Unix.
Because of the transition to `dirs` the global configuration file in
`/etc/xdg` was not picked up properly. With this commit if no config
file is found in the `${XDG_CONFIG_HOME}`, `nvfancontrol` will scan the
`${XDG_CONFIG_DIRS}` for the configuration file before resorting to the
built-in curve. Fixes #10.