Account for multipart driver versions when checking compatibility
This fixes a bug where `nvfancontrol` expects the driver version to
always be in the format "XXX.YY". This is true for stable releases but
beta and experimental drivers usually have a third part "XXX.YY.ZZ" which
would lead to a crash when trying to parse it into a float. This is now
addressed by splitting the driver version on the dots "." and keeping only
the first two parts (or "00" if only one part is found).