[Windows] Add functions for cooler info using the NVAPI client API
This commit adds some partial interface to the `ClientFan*` family
of functions introduced with recent NVAPIs. It's required for newer
GPUs to be able to report and set cooler statuses. As this is still
missing the `ClientFanCooler{Set,Get}Control` functions it's not
wired to the frontend functions of `nvctrl` just yet. Additional
testing is required to see if these functions are actually doing
anything on GPUs older than the 20xx series.