Never touch the fans when in monitor mode
By default the Drop trait of NVFanManager will reset control of the
fans back to automatic. This can cause a problem with monitor mode as the
fan curve itself might be managed from a third-party tool. By resetting
the fan back to automatic this will force the third-party tool to be
set back to automatic mode and therefore follow the GPU firmware when it
comes to fan management. This commit moves the monitor functionality into
NVFanManager and Drop will only reset to automatic only if monitor mode is
not enabled.